azure-functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

./local_wheels/prophet-1.0.1-py3-none-any.whl
./local_wheels/PyMeeus-0.5.11-py3-none-any.whl
azure-functions
azure-data-tables
numpy
pandas
--extra-index-url https://download.pytorch.org/whl/cpu
torch